pyproject.toml
[tool.poe.tasks]
make         = "python manage.py makemessages -l fa"
compile      = "python manage.py compilemessages"         # a simple command task
serve.script = "my_app.service:run(debug=True)"              # python script based task
tunnel.shell = "ssh -N -L 0.0.0.0:8080:$PROD:8080 $PROD &" 

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "crm.settings"


[tool.poetry]
name = "crm motor"
version = "2.1.1"
description = "orders managment and crm."
authors = ["nobody"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.5"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
arabic-reshaper = "3.0.0"
arrow = "1.2.3"
asgiref = "3.7.2"
async-timeout = "4.0.2"
asynctest = "0.13.0"
attrs = "23.1.0"
backports-zoneinfo = "0.2.1"
binaryornot = "0.4.4"
bleach = "6.0.0"
cached-property = "1.5.2"
certifi = "2023.5.7"
chardet = "5.1.0"
charset-normalizer = "3.1.0"
click = "8.1.3"
cookiecutter = "1.7.3"
crispy-tailwind = "0.5.0"
django = "3.2.19"
django-crispy-forms = "2.0"
django-environ = "0.10.0"
django-generic-filters = "2.1.1"
django-jalali-date = "1.0.2"
django-phonenumber-field = "7.1.0"
django-tailwind = "2.2.2"
djangorestframework = "3.14.0"
frozenlist = "1.3.3"
future = "0.18.3"
gunicorn = "20.1.0"
idna = "3.4"
importlib-metadata = "6.6.0"
isodate = "0.6.1"
jdatetime = "4.1.1"
jinja2 = "3.1.2"
jinja2-time = "0.2.0"
lxml = "4.9.2"
markdown = "3.4.3"
markupsafe = "2.1.3"
multidict = "6.0.4"
munch = "3.0.0"
packaging = "23.1"
pillow = "9.5.0"
platformdirs = "3.5.1"
poyo = "0.5.0"
pyparsing = "3.0.9"
python-bidi = "0.4.2"
python-dateutil = "2.8.2"
python-slugify = "8.0.1"
pytz = "2023.3"
reportlab = "4.0.4"
requests = "2.31.0"
requests-file = "1.5.1"
requests-toolbelt = "1.0.0"
six = "1.16.0"
sqlparse = "0.4.4"
text-unidecode = "1.3"
typing-extensions = "4.6.3"
urllib3 = "2.0.3"
webencodings = "0.5.1"
whitenoise = "6.4.0"
yarl = "1.9.2"
zeep = "4.2.1"
zipp = "3.15.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
django-admin-interface = "^0.26.0"
poethepoet = "^0.21.1"
phonenumbers = "^8.13.18"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
